V.0.0.1.43 [PRE-ALPHA] Diamond2DGL on process: Multiple Problems -> 1 - Bad Physics runtime 2 - Bad View Port

+++ src/main/java/diamond2DGL/EntityFactory.java
@@ -14,7 +14,7 @@ public class EntityFactory {
         return e;
     }
 
-    public static Entity createSpriteEntity(Sprite sprite, int sizeX, int sizeY) {
+    public static Entity createSpriteEntity(Sprite sprite, float sizeX, float sizeY) {
         Entity entity = createEntity("Sprite_Entity");
         entity.transform.scale.x = sizeX;
         entity.transform.scale.y = sizeY;
